pyproject.toml
[project]
name = "image-description-service"
version = "0.1.0"
description = ""
authors = [
    {name = "0xN1ck N1ckNam3",email = "0xn1ck@sova.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "langdetect (>=1.0.9,<2.0.0)",
    "accelerate (>=1.4.0,<2.0.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "qwen-vl-utils (>=0.0.10,<0.0.11)",
    "bitsandbytes (>=0.45.2,<0.46.0)",
    "torchvision (==0.21.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "lxml (>=5.3.1,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
torch = "^2.6.0"

